[UPD] ProfileViewModel
