+++ BestRecipes/Services/DataManager.swift
@@ -74,6 +74,12 @@ enum MealTypes: String, CaseIterable {
     }
 }
 
+enum SavedRecipesType: String {
+    case mine = "myRecipes"
+    case favorites = "favoritesRecipes"
+    case recent = "recentRecipes"
+}
+
 final class DataManager {
     static let shared = DataManager()
     
@@ -150,6 +156,37 @@ final class DataManager {
         } catch {
             print("Ошибка загрузки картинки: \(error)")
             return Data()
+        }
+    }
+    
+    func getRecipesFrom(_ storage: SavedRecipesType) -> [Recipe] {
+        if let recipesData = UserDefaults.standard.data(forKey: storage.rawValue),
+           let recipes = try? JSONDecoder().decode([Recipe].self, from: recipesData) {
+            return recipes
+        }
+        return []
+    }
+    
+    func addRecipe(_ recipe: Recipe, to storage: SavedRecipesType) {
+        var recipes = getRecipesFrom(storage)
+        
+        if recipes.contains(where: { $0.id == recipe.id }) { return }
+        
+        recipes.append(recipe)
+        
+        if let item = try? JSONEncoder().encode(recipes) {
+            UserDefaults.standard.set(item, forKey: storage.rawValue)
+        }
+    }
+    
+    func deleteRecipe(_ recipe: Recipe, from storage: SavedRecipesType) {
+        var recipes = getRecipesFrom(storage)
+        
+        guard let index = recipes.firstIndex(where: { $0.id == recipe.id }) else { return }
+        recipes.remove(at: index)
+        
+        if let item = try? JSONEncoder().encode(recipes) {
+            UserDefaults.standard.set(item, forKey: storage.rawValue)
         }
     }
 }
